refactor & chore
