+++ .trunk/trunk.yaml
@@ -1,0 +1,11 @@
+version: 0.1
+cli:
+  version: 0.12.1-beta
+lint:
+  enabled:
+    - eslint@8.10.0
+    - gitleaks@8.3.0
+    - markdownlint@0.31.1
+    - prettier@2.5.1
+    - stylelint@14.6.1
+    - stylelint-fmt@14.6.1
