+++ src/components/Layout/index.tsx
@@ -1,4 +1,4 @@
-import { PropsWithChildren } from 'react';
+import type { PropsWithChildren } from 'react';
 
 import GlobalNav from '../GlobalNav';
 import SideNav from '../SideNav';
