+++ src/components/_common/Loading/styles.module.scss
@@ -1,8 +1,8 @@
 .container {
   display: flex;
   flex-direction: column;
-  justify-content: center;
   align-items: center;
+  justify-content: center;
   width: 100vh;
   height: 60vh;
 }
