+++ src/pages/dashboard/_components/AdStatus/StatusChart/dropDownCategories.ts
