+++ src/pages/dashboard/_components/AdStatus/style.module.scss
@@ -35,7 +35,6 @@
       li {
         box-sizing: border-box;
         display: flex;
-        gap: 146px;
         align-items: flex-end;
         justify-content: space-between;
         width: 304px;
@@ -46,6 +45,7 @@
         border-radius: 10px;
 
         .leftBox {
+          margin-left: 25px;
           font-style: normal;
           white-space: nowrap;
 
@@ -71,6 +71,7 @@
 
         .rightBox {
           display: flex;
+          margin-right: 10px;
 
           svg {
             width: 10px;
@@ -108,6 +109,9 @@
       }
 
       .menu {
+        position: relative;
+        top: 7px;
+        right: 5px;
         width: 123px;
         height: 40px;
         background: colors.$BG_W;
