+++ src/pages/dashboard/_components/DatePicker/style.module.scss
@@ -31,8 +31,8 @@
 
   .calenderWrapper {
     position: absolute;
-    right: 10px;
-    bottom: -250px;
+    right: 25px;
+    bottom: -260px;
     z-index: levels.$DROPDOWN;
     width: 235px;
     height: 245px;
@@ -124,6 +124,15 @@
 
         &:hover {
           background-color: colors.$ERROR;
+        }
+      }
+
+      .noData {
+        color: colors.$GREY_50;
+        cursor: auto;
+
+        &:hover {
+          background-color: colors.$BG_W;
         }
       }
