+++ src/pages/dashboard/_states/compareDatesAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const compareDatesAtom = atom<string[]>({
+  key: 'compareDatesAtom',
+  default: [],
+});
