+++ src/pages/dashboard/_states/dayOrWeeklyAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const dayOrWeeklyAtom = atom<boolean>({
+  key: 'dayOrWeekAtom',
+  default: false,
+});
