+++ src/pages/dashboard/_states/mainIdxAtom.ts
@@ -4,8 +4,3 @@ export const mainIdxAtom = atom({
   key: 'mainIdxAtom',
   default: 0,
 });
-
-export const subIdxAtom = atom({
-  key: 'subIdxAtom',
-  default: -1,
-});
