+++ src/pages/dashboard/_states/mediaGridAttributesAtom.ts
@@ -10,20 +10,28 @@ const DEFAULT: MediaDataDisplayMap[] = [
     display: '매출',
   },
   {
+    key: 'roas',
+    display: 'ROAS',
+  },
+  {
     key: 'imp',
-    display: '노출 수',
+    display: '노출수',
   },
   {
     key: 'click',
     display: '클릭 수',
   },
   {
-    key: 'convValue',
-    display: '전환 수',
+    key: 'ctr',
+    display: '클릭률 (CTR)',
+  },
+  {
+    key: 'cpc',
+    display: '클릭당비용 (CPC)',
   },
 ];
 
-export const mediaChartAttributesState = atom<MediaDataDisplayMap[]>({
-  key: '#mediaChartAttributesState',
+export const mediaGridAttributesAtom = atom<MediaDataDisplayMap[]>({
+  key: '#mediaGridAttributesAtom',
   default: DEFAULT,
 });
