+++ src/pages/dashboard/_states/subIdxAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const subIdxAtom = atom({
+  key: 'subIdxAtom',
+  default: -1,
+});
