+++ src/pages/dashboard/_states/weelkyAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const weeklyAtom = atom<string[]>({
+  key: 'weeklyAtom',
+  default: [],
+});
