+++ src/pages/manage/_components/AdManageList/AdManageCard.tsx
@@ -1,22 +1,18 @@
-import { useState, useEffect } from 'react';
-import { useSetRecoilState } from 'recoil';
+import { useState } from 'react';
+import { useRecoilState } from 'recoil';
 import cx from 'classnames';
 
 import type { MouseEventHandler } from 'react';
-import { dataListAtom } from 'pages/manage/_states/dataList';
+import { dataListAtom } from 'pages/manage/_states/dataListAtom';
 import styles from './style.module.scss';
 
 const AdManageCard = ({ card, onClickSelectCard }: AdManageCardProps) => {
-  const setDataList = useSetRecoilState(dataListAtom);
+  // const [, setDataList] = useRecoilState(dataListAtom);
   const [isEditing, setIsEditing] = useState(false);
 
   const handleClickEdit = () => {
     setIsEditing((prev) => !prev);
   };
-
-  useEffect(() => {
-    console.log(isEditing);
-  }, [isEditing]);
 
   return (
     <article
