+++ src/pages/manage/_components/AdManageList/style.module.scss
@@ -35,15 +35,17 @@
       .formTitle {
         height: 19px;
         margin-bottom: 40px;
-        @include font.font('Roboto', '16', '700');
+        font-size: 19px;
+        font-weight: 700;
         line-height: 19px;
-        color: #3a474e;
+        color: colors.$GREY_800;
         user-select: none;
       }
 
       .table {
         width: 265px;
         height: 240px;
+        font-size: 14px;
 
         .category {
           position: relative;
@@ -51,6 +53,8 @@
           align-items: center;
           width: 100%;
           height: 40px;
+          font-size: 12px;
+          font-weight: 700;
           line-height: 40px;
 
           &::after {
@@ -68,17 +72,17 @@
             width: 40%;
             height: 14px;
             line-height: 14px;
-            @include font.font('Roboto', '12', '500');
-            color: #94a2ad;
+            color: colors.$GREY_300;
             user-select: none;
           }
 
           input {
             width: 60%;
             height: 14px;
-            @include font.font('Roboto', '12', '700');
+            font-size: 14px;
+            font-weight: 700;
             line-height: 14px;
-            color: #3a474e;
+            color: colors.$GREY_800;
             letter-spacing: 0.01em;
             cursor: pointer;
           }
@@ -87,9 +91,8 @@
             position: absolute;
             right: 5%;
             height: 14px;
-            @include font.font('Roboto', '12', '700');
             line-height: 14px;
-            color: #ee1e1e;
+            color: colors.$ERROR;
           }
         }
       }
@@ -107,9 +110,10 @@
 
         .editText {
           height: 16px;
-          @include font.font('Roboto', '14', '700');
+          font-size: 14px;
+          font-weight: 700;
           line-height: 16px;
-          color: #3a474e;
+          color: colors.$GREY_800;
         }
       }
     }
