+++ src/pages/manage/_utils/data.ts
@@ -1,4 +1,4 @@
-import AD_DATA from './wanted_FE_ad-list-data-set.json';
+import AD_DATA from 'data/wanted_FE_ad-list-data-set.json';
 
 const { ads } = AD_DATA;
