+++ src/styles/base/_fonts.scss
@@ -1,10 +1,11 @@
 @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Roboto&family=Stylish&display=swap');
 
 body,
 button,
 input,
 textarea {
-  font-family: Inter, Arial, PingFangSC-Regular, 'Microsoft YaHei', sans-serif;
+  font-family: Roboto, Inter, Arial, PingFangSC-Regular, 'Microsoft YaHei', sans-serif;
   font-feature-settings: 'tnum';
 }
