+++ src/styles/mixins/_font.scss
@@ -9,6 +9,7 @@ $FONT_SIZE_MAP: (
   '14': 14px,
   '12': 12px,
 );
+
 $FONT_WEIGHT_MAP: (
   '900': 900,
   '800': 800,
