--- src/types/adList.d.ts
@@ -1,21 +1,0 @@
-interface IAdList {
-  count: number;
-  ads: Ad[];
-}
-
-interface Ad {
-  id: number;
-  adType: string;
-  title: string;
-  budget: number;
-  status: string;
-  startDate: string;
-  endDate?: string;
-  report: Report;
-}
-
-interface Report {
-  cost: number;
-  convValue: number;
-  roas: number;
-}
