+++ src/types/manage.d.ts
@@ -20,3 +20,20 @@ interface AdManageDataListType {
   ads: AdManageFormItemsType[];
   count: number;
 }
+
+interface Ad {
+  id: number;
+  adType: string;
+  title: string;
+  budget: number;
+  status: string;
+  startDate: string;
+  endDate?: string;
+  report: Report;
+}
+
+interface Report {
+  cost: number;
+  convValue: number;
+  roas: number;
+}
