+++ src/types/trend.d.ts
